GAN
This code make machine generate anime faces with GAN.
It(Generator) first generate some random anime faces(complex Distribution) depending on the database(normal Distribution).
Then it(Discriminator) will rate the pictures on a scale of 0 to 1.
In the end, it(dcGAN) will keep training Generator and Discriminator.
You can find that the pictures it generates will more and more like a real  anime face instead of wired color block.